updated value storage edit button
On the 'Confirm.js' page when user clicks on 'Edit' button they will get taken to 'Amount.js' where the previously entered value will be stored in the input field. Same for date & time, and address.

When user clicks on 'Confirm' button on the 'Confirm.js' page, the values will no longer be stored and they get blank pages to make a new donation.